Explicitly cast numeric literals to type byte to avoid ESP32 and ESP8266 compile errors

+++ src/SerLCD.cpp
@@ -337,15 +337,15 @@ void SerLCD::home()
  * column - byte 0 to 19
  * row - byte 0 to 3
  *
- * returns: boolean true if cursor set.
  */
 void SerLCD::setCursor(byte col, byte row)
 {
   int row_offsets[] = {0x00, 0x40, 0x14, 0x54};
 
   //kepp variables in bounds
-  row = max(0, row);            //row cannot be less than 0
-  row = min(row, MAX_ROWS - 1); //row cannot be greater than max rows
+  //Explicitly cast numeric literals to type byte to avoid ESP32 and ESP8266 compile errors
+  row = max((byte) 0, row);            //row cannot be less than 0
+  row = min(row, (byte)(MAX_ROWS - 1)); //row cannot be greater than max rows
 
   //send the command
   specialCommand(LCD_SETDDRAMADDR | (col + row_offsets[row]));
